refactor: FetchRes로 공통타입 이름 변경
