+++ src/fetcher/backendFetch.ts
@@ -1,7 +1,8 @@
 import httpClient from '@/fetcher/fetch.ts';
 import * as process from 'process';
-import { isServer } from '@/fetcher/index.ts';
+
 import { responseInterceptor } from '@/fetcher/interceptors.ts';
+import { isServer } from '@/utils';
 
 // NOTE: 별도의 쿠키 조작을 하지 않는 기본 fetch 입니다. backend api에 request를 전송합니다.
 // client side 에서 사용할 수 있습니다.
